fix: extract version mismatch in new component
